improve request timeout handling
This commit improves the handler timeout
implementation such that:
 - No more sporadic data races (concurrent map writes) occur
 - Panics in the actual HTTP handler no longer get propagated,
   and therefore, shadowed by the timeout handler.

Per handler timeouts are necessary since some handler (i.e.
log tracing) should not time out while others, like key generation,
should time out after a short period of time.

However, the Go HTTP server implementation only supports global
timeouts.

Therefore, this commit has apply some tricks to start an HTTP handler
in a separate Go routine and safely abort it once the timeout exceeds.

One major problem is that the timeout handler has to recover from a
potential panic within the actual HTTP handler and abort the request
without shadowing the panic message nor its original location.

Therefore, this implementation reuses some of the techniques used
by the Go standard library. It spawns the handler in a new go routine
and, in case of a panic, logs it to the server's error log and then
panics `http.ErrAbortHandler`.
This error is handled by the HTTP server specially in the sense that
it aborts the request but does not log an error message such that we
don't end up with two error log messages for one panic.

Further, the new timeout applies to the entire lifetime of the request
and does not get reset once the handler writes to the client connection.
So it behaves more like the global server timeouts but on a
handler-basis.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>