proxy: avoid passing nil as http.ResponseWriter to handlers (#140)
This commit fixes a bug in the `TLSProxy` handler function
that occurs when KES does not create an audit log entry for
the request. This happens only in case of `/v1/metrics` API.

In case of a `/v1/metrics` API request, KES will not create
an `AuditResponseWriter`. In this case, a type check:
`aw, ok := w.(*AuditResponseWriter)` will result in `aw == nil`.

Therefore, not the assumed `AuditResponseWriter` but the actual
`http.ResponseWriter` must be passed to any subsequent handler.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>