secret: add key ID to encrypted keys
This commit adds unique key IDs to generated
ciphertexts. The key ID identifies the root/master
key used to generate the ciphertext. Now it is possible,
to determine whether a particular key was used to encrypt
a certain ciphertext.

However, the key ID is just for informational purposes.
It is derived from the root/master key content as:
```
id := HEX(SHA-256(key)[:16])
```

The ID is a truncated hash for two reasons:
 1. Avoid additional size overhead
 2. Limit the damage a broken hash function (i.e. SHA-256)
    could cause.

W.r.t. truncating SHA-256 to 128 bits:
This is not a security issue but instead a common technique.
In particular, SHA-224 is truncated SHA-256 and SHA-384 is truncated
SHA-512.
Truncating the hash only reduces the collision resistance boundary.
However, we are not interested in collision resistance w.r.t. an
attacker crafting a collision. Instead, we only want to use that
IDs are "sufficiently" unique - such that two keys have different
IDs with very high probability.

The most important property of the ID is that it must not reveal
anything about the root/master key content.

This should be the case since we use a collision resistant hash
function. Further, even if SHA-256 would reveal information about
the hashed message (and therefore would be crypto. broken) we
limit the damage by truncating the hash to 128 bits. It's very
unlikely that a 128 bit hash value would reveal more than 128 bit
about the message - here the hashed key.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>