config: rename root identity
This commit renames the root identity to
admin identity. Therefore, it changes
the config file structure:

Before:
```yaml

root: ...
```

After:
```yaml

admin:
  identity: ...
```

Further, this commit removes the `--root` CLI flag.
From now on, the admin identity has to be specified
in the config file - but it can be inferred from the
environment.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>