config: restructure server initialization and configuration
This commit restructures the server initialization code.
Therefore, configuration-related code (types, parsing,
(de)serialization) has been moved to an internal `yml`
package.

Further, this commit improves the AWS SecretsManager
and Hashicorp Vault implementation to be self-contained
and don't depend on default configuration values.

The goal of this commit is to decouple and stabilize
internal APIs - i.e. for a stateful KES server implementation.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>