testing: new package kestest
This commit introduces a new sub-package
`kestest` that contains utilities for
end-to-end KES testing.

It is heavily inspired by the `net/http/httptest`
package and contains a `Server` type that implements
a minimal KES server.

The new `kestest` package makes it possible to
write unit tests covering end-to-end usage scenarios
without a dedicated KES server for testing purposes.

Fixes #164

Signed-off-by: Andreas Auernhammer <hi@aead.dev>