status: expose backend KMS status
This commit changes the `/v1/status` API by:
 - adding a backend KMS status
 - removing the TLS certificate information

The KMS status is generic for all KMS backends
and describes whether the backend KMS is reachable
or unreachable as well as the network latency between
KES and the backend KMS instance.

The TLS certificate information got removed from the
status API since this information can be gathered by
the client itself (by looking at the TLS certificate).

Exposing this information in the status API is particularly
useful and just bloats the API.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>