cache: add offline caching of keys
This commit adds offline caching of keys.

KES keeps keys - fetched from the KMS key
store - in a in-memory cache to increase
performance and reduce the request rate
to the central KMS.
Entries in the cache expiry after a configurable
time period and get removed by a cache GC.

Usually it is recommended to keep the cache
expiry periods quite low - e.g.:
 - Any:    5m0s
 - Unused:  20s

In particular, low cache expiry values reduce
the time window KES can operate without interacting
with the central KMS when serving stateless
requests; e.g. generating a new data encryption key.
Note that KES can never server stateful requests,
like creating or deleting a key, without the KMS.

Especially in distributed setups, two KES servers
will sync eventually once their cache entires have
expired. For example, one KES server receives a
request to delete a key from the KMS key store.
The second KES server will not notice that this key
got deleted until its corresponding cache entry
has expired.
Low cache expiry values reduce the time window
when multiple KES servers are not synchronized.

However, low cache expiry values require that the
KMS is highly available. As soon as a cache entry
expires, KES needs to reach out to the KMS to
fetch the key again.
If the KMS is not available, KES will not be able
continue serving stateless requests.

It may be desirable to keep keys longer in the cache
to reduce the impact of the central KMS being down
and continue serving stateless requests - but only
when the KMS is actually down. When the KMS is
available, KES should expiry keys relatively quickly
and only cache them longer when the KMS is not reachable.

This commit adds this ability by another cache expiry
configuration:
```yaml
cache:
  expiry:
    any:       5m0s
    unused:     30s
    offline: 1h0m0s
```

Now, KES will cache keys for one hour if and only
if the KMS is not available. As soon as the KMS
is reachable again, KES clears the cache to sync
with the central KMS again.

If no `offline` expiry is set, KES will not cache
keys when the KMS is down. It will simply not
use an offline cache.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>

Co-authored-by: Klaus Post <klauspost@gmail.com>
Signed-off-by: Andreas Auernhammer <hi@aead.dev>