improve TLS private key / certificate parsing
This commit improves the TLS private key and
certificate parsing.

Now, a TLS private key file can contain multiple
PEM blocks - each being either a `PRIVATE KEY`
or a `CERTIFICATE`. It is now possible to bundle
the private key and certificate in a single
private key file. This may be useful for e.g. backups.

In contrast, a certificate file can contain multiple
PEM blocks -- each must be a `CERTIFICATE`.
Private keys are not allowed in a certificate file
b/c a certificate are public information and may
be copied to other machines for manual trust establishment.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>