metric: expose usable CPUs
This commit exposes the number of
usable CPUs - in addition to the
number of available CPU cores.

The number of available CPU cores
shows how many logical CPU cores
are present on the system.

However, the KES server may only
use a subset of these CPUs.
For example when `GOMAXPORCS` is set.

The number of usable CPU cores,
exposed by this commit, shows how
many logical CPU cores can be used
by the KES server.
It may be less then or equal to the
number of available CPUs.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>