cmd: display server endpoints listening on correctly (#229)
This commit fixes a UI bug when displaying the reachable
IP endpoints.

Before, the KES server always printed the IPv4 addresses of
all available network interfaces. This is incorrect and can
lead to confusion.

The KES server listens on the specified network interface(s).
This can be a specific network interface:
 - e.g. `192.168.188.162`
 - the pseudo address `0.0.0.0`

When listening on a specific IP / interface - e.g. `192.168.188.162`
then KES accepts connections only from that interface.
For example a request to `https://192.168.188.162:7373/version` will served
while a connection attempt to `https://127.0.0.1:7373/version` will be
refused.

In particular, listening on `:<port>` resp. `0.0.0.0:<port>` is not
the same as listening on `127.0.0.1:<port>` since the later listens
**only** on the loopback device.

Now, KES displays the endpoints / network interfaces it is listening on
correctly.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>