vault: issue health check before dialing
This commit fixes an issue in the Vault
health-check logic.

Before, KES always opened a TCP connection
to Vault. In case of TLS, this causes Vault
to log a TLS handshake error for every KES
status check.

Now, KES only tries to check if Vault is reachable
at all if the regular health-check fails.
So, we have run the health check in reverse order.

Fixes #230

Signed-off-by: Andreas Auernhammer <hi@aead.dev>