dep: update `gopkg.in/yaml` to `v3`
This commit updates the YAML parser to `v3`.
This fixes a depend-a-bot alert for `v2`:

```
An issue in the Unmarshal function in Go-Yaml v3 causes
the program to crash when attempting to deserialize invalid
input.
```

The new `v3` YAML parser works as before. It successfully parses
a YAML config with all possible configuration fields and also replaces
env. variable references.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>