improve timer usage by re-using them
timers can trickle memory alloc quite
sharply over a period of time for days.

Since the memory build up is so slow,
the buildup is not easily visible,
additionally there is a cost to always
invoking runtime timer in a loop.

This PR is an attempt to address this
by reusing the timer by creating it
once and making sure to use the right
amount of "Reset()" to wait for expected
lengths.

Additionally fixes a timer leak in
ReloadAfter() certificate reload function.