separate stateful and stateless server implementation
This commit separates the stateful and stateless server (gateway)
implementation.

In particular, the abstracts for the gateway (e.g. KMS key store)
and the stateful server cause additional complexity and often
leak the underlying implementation or are fundamentally incorrect.

For example, the gateway should be stateless and read-only (except
for the managed keys) while the server require modifications to
policies, identities etc.

Therefore, mid and long-term it seems better to fundamentally separate
the gateway and server API such that each can be implemented specific
for the requirements.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>