cache: fix bug in offline caching
This commit fixes a bug in the offline
caching. Before, the `unused` caching time
was used for the offline cache - which is incorrect.

This leads to confusing offline caching behavior.
The offline cache works for some amount of time,
precisely the `unused` time.

This commit fixes this by using the correct `offline`
time.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>