cmd: reject server certificates without SANs
This commit disallows the use of TLS server
certificates that do not contain any SANs.

Certificate verification fails when a certificate
contains a CN but not a SAN since Go 1.15.
Ref: https://go.dev/doc/go1.15#commonname

We don't reject client certificates without a SAN
since client certificates may not be verified at
all due to the identity verification.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>