keserv: add version field to server config
This commit adds a version field to the
server configuration. It's purpose is to
reliably detect and parse incompatible
server config files in the future.

Currently, the only valid value for the
version field is "v1" and the field itself
is optional - for backward compatibility.
```
version: v1
address: 0.0.0.0:7373
...
```
Future versions of KES will eventually require
this field.

Based on the version field we can now distinguish
incompatible config files and parse or reject them
with a proper messaging.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>