add `KeyAlgorithm` enum type
This commit adds a enum for cryptographic
algorithms.

In general, a crypto. key is specific for
one crypto. algorithm. This commit exposes
the crypto. algorithm as enum.

Further, this commit changes the key data type
such that its binary representation is more
compact by using the `KeyAlgorithm` enum
instead of a string type.

In addition, this commit adds a version
identifier to the binary and text representation
of keys that allows easier backwards-compatible
changes to the key data type.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>