move from UBI-minimal to UBI-micro containers
This commit replaces the UBI-minimal container image
with a multi-level docker build for a UBI-micro
container.

This reduces the image size and avoids unneccesary
bloat - which may be flagged by security scanners.

A multi-level docker build is necessary because
the UBI-micro image does not contain a package
manager and we need to update the CA certificate
bundle.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>