gcp: use default GCP endpoint for status checks
This commit fixes a bug in the GCP SecretsManager
status check implementation.

Commit 64f37b1 changed the status check for a general
TCP status check to a keystore specific HTTP status
check.
As part of this change, the GCP implementation
only considered the SecretsManager endpoint
provided in the config file - which is usually
not set - and did not fallback to the GCP default:
`secretmanager.googleapis.com:443`

This commit fixes this such that it is no longer
required to set the GCP endpoint to successfully
query the SecretsManager liveness.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>