add readiness API
This commit adds a new API to check the readiness
of a KES edge server.

In general, liveness and readiness may be checked
by a system like Kubernetes to determine whether
a system (i.e. pod) is available and ready to serve
requests.

In general, the liveness probe should report
whether a system is not in a deadlock state.
In particular, it does not indicate whether
a system is "alive".

In contrast, the readiness probe should report
whether a system is ready to handle requests.
For example, if KES cannot reach its backend
keystore its readiness probe should fail since
it will not be able to process (some) requests.

For KES the liveness probe makes no sense at
the moment since it is not a distributed systems
where deadlocks can occur. Hence, it is not
implemented.

To make readiness really useful we have to change
the KES TLS behavior in a subtle way.

By default, KES doesn't not accept a TLS connection
attempt from a client without a TLS certificate
(or API key that translates to a certificate).
Hence, only clients with an API key or certificate
can check the KES liveness or readiness.

However, some systems (e.g. K8S) do not support
client certificates for liveness and readiness
probes. Hence, we allow the disabling authentication
for these two APIs in the `api` section in the
KES config file.

To make this actually work, we also have to tell
the Go TLS stack to allow TLS handshakes without
a certificate. Therefore, whenever at least one
API has a `skip_auth` config set to true, we also
set:
- `tls.RequestClientCert` instead of `tls.RequireAnyClientCert`
- `tls.VerifyClientCertIfGiven` instead of `tlsRequireAndVerifyClientCert`

Ref to: https://pkg.go.dev/crypto/tls#ClientAuthType

Now, the subtle behavior change is the following:
If the config file disables authentication for any
API, clients are no longer forced to send a client
certificate whenever they try to perform an API call.
Instead, the TLS handshake will always succeed.
However, for all APIs that require authentication
the HTTP API handler will reject the request with
the following error:
```
no client certificate is present
```
The difference is the protocol layer that is responsible
for rejecting clients without certificates. Before,
and also today in the default config, it's the Gop TLS
stack rejecting such connection attempts. However,
when authentication is disabled for at least one API,
the TLS stack will allow requests without a client
certificate and its no the HTTP / API layer that
has to enforce that the client sends a client certificate
when the API requires authentication.

This however, does not change the security model.
APIs that required authentication still require
authentication.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>