kesconf: expose public API
This commit exposes kes server confguration API and
provides a dev server.

Now, the `github.com/minio/kes/kesconf` package exposes
a stable API for reading the KES server configuration.
This supersedes the `kv` package and removes the `edge`
package.

Further, the KES server provides the `--dev` flag to start
a KES server in development mode with a in-memory keystore.
The `--key`, `--cert` and `--auth` are deprecated. Users should
use the config file to provide such config options.

Signed-off-by: Andreas Auernhammer <github@aead.dev>