Parse address from config if not provided with '--addr' (#419)
This fixes a bug where kes would try to split and evaluate the listen
address, before the config file was read. This caused kes to fail if no
`--addr` value was provided.

With this change, the config file is parsed before we try to determine
the listen address. If the `--addr` parameter is provided, we override
the value from the config as expected.

Co-authored-by: Robert Lützner <robert.luetzner@iternity.com>