vault: support authentication in different namespaces
This commit adds support for authenticating within a different
namespace compared to the secret engine namespace.

For example, KES should use the AppRole authentication engine
within the root namespace to authenticate but then use the
K/V and transit engine within a user-specific namespace, like
'foo'. Similarly, KES may authenticate in the namespace 'foo'
but use the secret engines in namespace 'bar' or 'foo/bar'.
With this change, this use case is now supported.

The special namespace string '/' is treated as alias for the root
namespace and causes KES to not send any namespace header for
authentication. If the namespace header for root should be set
explicitly, use the namespace 'root/'.

Signed-off-by: Andreas Auernhammer <github@aead.dev>