keystore: use pre-configured client for {Gemalto,Fortanix} status check
This commit fixes an issue on the Gemalto and Fortanix status checks.
Before, the status check for both backends used the `http.DefaultClient`.
This can cause TLS verification issues when the backend uses certificate
not issued by one of the server's root CAs - e.g. with a custom root CA.

Now, the status check uses the HTTP client used for other API calls, like
creating a key, with a TLS configuration provided to the server.

Signed-off-by: Andreas Auernhammer <github@aead.dev>