vault: clone client TLS config and headers for status checks (#476)
This commit improves the Vault status check logic. KES has to
clone the Vault client since it has to clear the namespace header
for status checks.

However, KES should use the same TLS client config and token configured
on the client when performing such status checks. KES has to opt-in
to this behavior since the Vault SDK doesn't do this by default.

Signed-off-by: Andreas Auernhammer <github@aead.dev>